backend/requirements.txt
fastapi==0.115.0
pydantic[email]>=2.0.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
slowapi==0.1.9
